# openapi.yaml
openapi: 3.0.3
info:
  title: MonPetitBiz WhatsApp Bot API
  description: |
    A WhatsApp bot API for micro-businesses in Africa to record transactions, 
    track inventory, and generate financial reports through natural language conversations.
    
    ## Features
    - WhatsApp webhook integration
    - Natural language processing for business commands
    - Transaction recording (sales & expenses)
    - Inventory management
    - Financial report generation
    - PDF report delivery via WhatsApp
    - Comprehensive health monitoring
    
    ## Authentication
    Most endpoints require JWT authentication. Use the `/auth/send-otp` and 
    `/auth/verify-otp` endpoints to obtain an access token.
    
  version: 1.0.0
  contact:
    name: MonPetitBiz Support
    email: support@monpetitbiz.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.monpetitbiz.com
    description: Production server
...
paths:
...
  /auth/send-otp:
    post:
      tags:
        - Authentication
      summary: Send OTP to phone number
      description: Sends a one-time password to the specified phone number for authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - phoneNumber
              properties:
                phoneNumber:
                  type: string
                  pattern: '^\+[1-9]\d{1,14}$'
                  example: "+221123456789"
                  description: Phone number in international format
            examples:
              burkina:
                summary: Burkina Faso phone number
                value:
                  phoneNumber: "+226123456789"
              mali:
                summary: Mali phone number  
                value:
                  phoneNumber: "+223123456789"
      responses:
        '200':
          description: OTP sent successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "OTP sent successfully"
        '400':
          description: Invalid phone number format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/verify-otp:
    post:
      tags:
        - Authentication
      summary: Verify OTP and authenticate
      description: Verifies the OTP code and returns an access token if valid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - phoneNumber
                - code
              properties:
                phoneNumber:
                  type: string
                  pattern: '^\+[1-9]\d{1,14}$'
                  example: "+221123456789"
                code:
                  type: string
                  pattern: '^\d{6}$'
                  example: "123456"
                  description: 6-digit OTP code
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
                  accessToken:
                    type: string
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        '401':
          description: Invalid OTP code
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register new user and business (Legacy)
      description: Creates a new user account with associated business (Legacy endpoint - use specific registration endpoints instead)
      deprecated: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - phoneNumber
                - businessName
                - role
              properties:
                phoneNumber:
                  type: string
                  pattern: '^\+[1-9]\d{1,14}$'
                  example: "+221123456789"
                businessName:
                  type: string
                  minLength: 2
                  maxLength: 100
                  example: "Boutique Fatou"
                role:
                  type: string
                  enum: [owner, seller]
                  example: "owner"
                language:
                  type: string
                  enum: [fr, wo]
                  default: fr
                  example: "fr"
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '409':
          description: User already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/register-business-owner:
    post:
      tags:
        - Authentication
      summary: Register business owner
      description: |
        Creates a new business owner account with an automatically generated business code.
        The business code is used by employees to join the business during their registration.
        
        **Process:**
        1. Validates phone number and business information
        2. Creates new business with unique 6-character code
        3. Creates owner user account linked to the business
        4. Returns business code for sharing with employees
        
        **Business Code Format:**
        - 6 characters (A-Z, 2-9)
        - Excludes confusing characters (0/O, 1/I/L)
        - Unique across all businesses
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - phoneNumber
                - businessName
                - ownerName
              properties:
                phoneNumber:
                  type: string
                  pattern: '^\+[1-9]\d{1,14}$'
                  example: "+221123456789"
                  description: Phone number in international format
                businessName:
                  type: string
                  minLength: 2
                  maxLength: 100
                  example: "Boutique Fatou"
                  description: Name of the business
                ownerName:
                  type: string
                  minLength: 2
                  maxLength: 100
                  example: "Fabrice Ilboudo"
                  description: Full name of the business owner
                language:
                  type: string
                  enum: [fr, wo]
                  default: fr
                  example: "fr"
                  description: Preferred language for communication
            examples:
              burkina_business:
                summary: Burkina Faso business registration
                value:
                  phoneNumber: "+226123456789"
                  businessName: "Boutique Fatou"
                  ownerName: "Fabrice Ilboudo"
                  language: "fr"
              mali_business:
                summary: Mali business registration
                value:
                  phoneNumber: "+223987654321"
                  businessName: "Magasin Aminata"
                  ownerName: "Aminata Traoré"
                  language: "fr"
      responses:
        '201':
          description: Business owner registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
                  business:
                    $ref: '#/components/schemas/BusinessWithCode'
                  accessToken:
                    type: string
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                    description: JWT access token for authentication
                  message:
                    type: string
                    example: "Entreprise créée avec succès! Partagez le code ABC123 avec vos employés."
              examples:
                success_response:
                  summary: Successful registration
                  value:
                    user:
                      id: "123e4567-e89b-12d3-a456-426614174000"
                      phoneNumber: "+221123456789"
                      employeeName: "Fabrice Ilboudo"
                      role: "owner"
                      language: "fr"
                      isActive: true
                    business:
                      id: "456e7890-e89b-12d3-a456-426614174001"
                      name: "Boutique Fatou"
                      businessCode: "ABC123"
                      currency: "XOF"
                      timezone: "Africa/Dakar"
                    accessToken: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                    message: "Entreprise créée avec succès! Partagez le code ABC123 avec vos employés."
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalid_phone:
                  summary: Invalid phone number
                  value:
                    statusCode: 400
                    message: "Numéro de téléphone invalide. Format attendu: +221XXXXXXXXX"
                    error: "Bad Request"
                invalid_business_name:
                  summary: Invalid business name
                  value:
                    statusCode: 400
                    message: "Le nom de l'entreprise doit contenir entre 2 et 100 caractères"
                    error: "Bad Request"
        '409':
          description: Phone number already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                duplicate_phone:
                  summary: Phone number already exists
                  value:
                    statusCode: 409
                    message: "Ce numéro de téléphone est déjà enregistré"
                    error: "Conflict"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                server_error:
                  summary: Server error
                  value:
                    statusCode: 500
                    message: "Erreur lors de la création de l'entreprise. Veuillez réessayer."
                    error: "Internal Server Error"
...
  /auth/business/{code}:
    get:
      tags:
        - Authentication
      summary: Get business information by code
      description: |
        Retrieves business information using a business code for validation purposes.
        This endpoint is used by employees to verify their business code before registration.
        
        **Use Cases:**
        - Employee verification of business code during registration
        - Displaying business name for confirmation
        - Validating business code format and existence
      parameters:
        - name: code
          in: path
          required: true
          description: 6-character business code
          schema:
            type: string
            pattern: '^[A-Z2-9]{6}$'
            example: "ABC123"
          examples:
            valid_code:
              summary: Valid business code
              value: "ABC123"
            another_code:
              summary: Another valid code
              value: "XYZ789"
      responses:
        '200':
          description: Business information retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  business:
                    $ref: '#/components/schemas/BusinessInfo'
                  message:
                    type: string
                    example: "Entreprise trouvée: Boutique Fatou"
              examples:
                business_found:
                  summary: Business found
                  value:
                    business:
                      id: "456e7890-e89b-12d3-a456-426614174001"
                      name: "Boutique Fatou"
                      businessCode: "ABC123"
                      employeeCount: 3
                      currency: "XOF"
                      timezone: "Africa/Dakar"
                    message: "Entreprise trouvée: Boutique Fatou"
        '400':
          description: Invalid business code format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalid_format:
                  summary: Invalid code format
                  value:
                    statusCode: 400
                    message: "Format de code invalide. Attendu: 6 caractères (ex: ABC123)"
                    error: "Bad Request"
        '404':
          description: Business code not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                code_not_found:
                  summary: Code not found
                  value:
                    statusCode: 404
                    message: "Code d'entreprise inexistant. Vérifiez avec votre employeur."
                    error: "Not Found"

  # WhatsApp Endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        phoneNumber:
          type: string
          example: "+221123456789"
        businessId:
          type: string
          format: uuid
        employeeName:
          type: string
          example: "Fabrice Ilboudo"
          description: Display name for the employee/owner
        role:
          type: string
          enum: [owner, seller, manager]
          example: "owner"
          description: User role determining access permissions
        language:
          type: string
          enum: [fr, wo]
          example: "fr"
        isActive:
          type: boolean
          example: true
        business:
          $ref: '#/components/schemas/Business'
    Business:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "456e7890-e89b-12d3-a456-426614174001"
        name:
          type: string
          example: "Boutique Fatou"
          description: Business name
        currency:
          type: string
          example: "XOF"
          description: Business currency code
        timezone:
          type: string
          example: "Africa/Dakar"
          description: Business timezone
    BusinessWithCode:
      allOf:
        - $ref: '#/components/schemas/Business'
        - type: object
          properties:
            businessCode:
              type: string
              pattern: '^[A-Z2-9]{6}$'
              example: "ABC123"
              description: Unique 6-character business code for employee invitations
    BusinessInfo:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "456e7890-e89b-12d3-a456-426614174001"
        name:
          type: string
          example: "Boutique Fatou"
          description: Business name
        businessCode:
          type: string
          pattern: '^[A-Z2-9]{6}$'
          example: "ABC123"
          description: Business code
        employeeCount:
          type: integer
          example: 3
          description: Number of employees in the business
        currency:
          type: string
          example: "XOF"
          description: Business currency code
        timezone:
          type: string
          example: "Africa/Dakar"
          description: Business timezone
...
    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: number
          example: 400
        message:
          type: string
          example: "Invalid request"
        error:
          type: string
          example: "Bad Request"
        timestamp:
          type: string
          format: date-time
...